requirements.txt
einops==0.3.2
ignite==1.1.0
matplotlib==3.4.3
MedPy==0.4.0
mmcv==1.3.16
monai==0.7.0
numpy==1.21.4
open3d==0.13.0
pytorch_ignite==0.4.4
pyvista==0.32.1
PyYAML==6.0
scikit_image==0.18.3
scipy==1.7.2
skimage==0.0
tensorboardX==2.4
timm==0.4.12
torch==1.12.0+cu116
torch-scatter==2.1.0+pt112cu116
torchvision==0.13.0+cu116
